Move CIRunner into CI package
